requirements.txt
streamlit
pillow
tensorflow